% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBoCo.R
\name{boxcoxnc2}
\alias{boxcoxnc2}
\title{Title}
\usage{
boxcoxnc2(i = "data to transform", data, lam = seq(-2, 2, 0.01),
  plotit = TRUE, p.method = "BY")
}
\arguments{
\item{i}{optional name of the vector which is being transform}

\item{data}{the vector to transform}

\item{lam}{the sequence of lambda to try}

\item{plotit}{logical}

\item{method}{which method should be used. Default to "all"}
}
\value{
a numeric vector transform using boxcox transformation
}
\description{
Title
}
\examples{
x <- rnorm(100)
boxcoxnc2(data =exp(x),lam = seq(-2, 2, 0.5))
}
